Add support for HTML4 downtransform
Tweaks now support a new tag, <compatibility>. It supports one
attribute, namely "html4-transform", which can be a path to a XSLT which
transforms HTML5 to HTML4.

If this is set, the information about the User-Agent will be used to
determine whether it supports HTML5. If that is not the case, the HTML4
downtransfrom will be applied to the result tree.