Add transformation of py:content attribute
py:content will now be handled like href attributes on certain elements,
but this will happen on _all_ elements in the tree supplied to
transformPyNamespace. It will be changed to a namespaces-less content
attribute (overriding any existing namespace-less content attributes).

If an element with py:content attribute also has a py:content-make-uri
attribute which is set to a boolean True value, the created content
attribute will hold the absolute URI (including the https? prefix).