Implemented Accept-Charset handling
The Context is now able to use the best character set for a particular
message. It goes through the preferred encodings of the client (limited
to five attempts, to prevent DoS) and tries to encode the result message
with each. If none of the encodings can encode the whole message, a
fallback to utf-8 is used. In any case, the Content-Type header will
reflect the correct encoding.