Fix missing charset attribute in Content-Type headers
In cases where the User-Agent did not send an Accept-Charset header, we
did not send a charset attribute back due to a bug in the
CharsetPreferenceList code.

The code made two mistakes:
(1) It accepted empty header values (equivalent to non-existing header
for Accept-Charset, as an empty header would be invalid).
(2) It did not insert a *;q=1.0, if no other values were present.