Add crumb class: StaticCrumb
Allows loading the contents of a document (parsed using one of the
loaded parsers) and render it as a crumb. Syntax:

   <static:crumb src="path/to/doc" type="mime/type" />

@type can be omitted, which makes PyXWF try to guess the mime-type based
on the files extension (this won't work for PyWebXML documents).

For the future, internal caching support is planned, which should
increase performance and reduce load on the global cache at the cost of
some memory. The memory used by this method will be _less_ than if the
global cache was used, but it will be a persistent use.